WIP: First implementation of EpiOffset.
This commit provides a simple class to help with coordinate calculations
when addressing pixels from an Epi. It is intended to simplify
coordinate arithmetic when iterating over Epi coordinates while
accessing multiple channels.

Note that the coordinate calculations are wrong! They need to be changed
to accommodate VIGRA's coordinate order.